move debug functions into their own file
