+++ exam_graph/debug.py
@@ -1,0 +1,22 @@
+from django.http import HttpResponse, JsonResponse, HttpRequest
+import pprint
+
+# Helper Debug Functions
+
+
+def debug_request(request:HttpRequest):
+    request_data = dir(request)  # Lists all attributes and methods of the request object
+    return pprint.pformat(request_data, indent=2)
+
+
+def prettify_request(request:HttpRequest):
+    return {
+        "method": request.method,
+        "headers": dict(request.headers),
+        "GET_params": dict(request.GET) if request.GET else None,
+        "POST_params": dict(request.POST),
+        "FILES": {k: str(v) for k, v in request.FILES.items()},
+        "path": request.path,
+        "content_type": request.content_type,
+        "session_keys": list(request.session.keys()),
+    }
