+++ exam_graph/views.py
@@ -4,7 +4,6 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse, JsonResponse, HttpRequest
 import pandas as pd
-import pprint
 from . import helper
 from . import filters
 import traceback
@@ -163,26 +162,6 @@ def filter_submission_handler(request):
             stack_trace = traceback.format_exc()  # Capture the full traceback
             print(stack_trace)  # Log the detailed error in the console
             return HttpResponse(f"{error_message}<br><pre>{stack_trace}</pre>", content_type="text/html")
-
-
-
-# Debugging functions:
-def debug_request(request:HttpRequest):
-    request_data = dir(request)  # Lists all attributes and methods of the request object
-    return pprint.pformat(request_data, indent=2)
-
-
-def prettify_request(request:HttpRequest):
-    return {
-        "method": request.method,
-        "headers": dict(request.headers),
-        "GET_params": dict(request.GET) if request.GET else None,
-        "POST_params": dict(request.POST),
-        "FILES": {k: str(v) for k, v in request.FILES.items()},
-        "path": request.path,
-        "content_type": request.content_type,
-        "session_keys": list(request.session.keys()),
-    }
 
 
 def load_data(request:HttpRequest):
